src/requirements.txt
azure-functions
requests
pandas